requirements.txt
# ========================================
# ANDROID APP AUTOMATION - DEPENDENCIES
# ========================================
# Framework: Appium-Python-Client + BrowserStack
# Python Version: 3.10+
#
# Installation:
#   pip install -r requirements.txt
# ========================================

# Core Automation Framework
Appium-Python-Client>=3.0.0
selenium>=4.15.0

# Utilities
python-dateutil>=2.8.2
